requirements.txt
flask
flask-mail
mysql-connector-python
itsdangerous
